+++ pyproject.toml
@@ -1,0 +1,7 @@
+[build-system]
+requires = [
+    "setuptools>=40.6.2",
+    "wheel",
+    "unasync==0.4.*",
+]
+build-backend = "setuptools.build_meta"
