+++ requirements.txt
@@ -23,5 +23,6 @@ trio
 trustme
 uvicorn
 seed-isort-config
+wheel
 
 attrs>=19.2  # See: https://github.com/encode/httpx/pull/566#issuecomment-559862665
