+++ scripts/clean
@@ -3,6 +3,9 @@
 if [ -d 'dist' ] ; then
     rm -r dist
 fi
+if [ -d 'build' ] ; then
+    rm -r build
+fi
 if [ -d 'site' ] ; then
     rm -r site
 fi
